+++ Editor/Config/ElympicsConfigEditor.cs
@@ -7,6 +7,7 @@ using Object = UnityEngine.Object;
 
 namespace Elympics
 {
+	[InitializeOnLoad]
 	[CustomEditor(typeof(ElympicsConfig))]
 	public class ElympicsConfigEditor : Editor
 	{
@@ -18,14 +19,15 @@ namespace Elympics
 		private SerializedProperty _elympicsApiEndpoint;
 		private SerializedProperty _elympicsLobbyEndpoint;
 		private SerializedProperty _elympicsGameServersEndpoint;
-		private SerializedProperty _elympicVersion;
-		private string _cachedSdkVersion = string.Empty;
+		private static string _cachedSdkVersion = string.Empty;
 
+		static ElympicsConfigEditor()
+		{
+			_cachedSdkVersion = ElympicsVersionRetriever.GetVersionStringFromAssembly();
+		}
 
 		private void OnEnable()
 		{
-			CheckSdkVersion();
-			_cachedSdkVersion = _elympicVersion.stringValue;
 			_elympicsApiEndpoint = serializedObject.FindProperty("elympicsApiEndpoint");
 			_elympicsLobbyEndpoint = serializedObject.FindProperty("elympicsLobbyEndpoint");
 			_elympicsGameServersEndpoint = serializedObject.FindProperty("elympicsGameServersEndpoint");
@@ -37,22 +39,7 @@ namespace Elympics
 
 		private void OnValidate()
 		{
-			CheckSdkVersion();
-			_cachedSdkVersion = _elympicVersion.stringValue;
-		}
-
-		private void CheckSdkVersion()
-		{
-			if (_elympicVersion == null)
-			{
-				_elympicVersion = serializedObject.FindProperty("elympicsVersion");
-			}
-
-			if (!EditorApplication.isPlayingOrWillChangePlaymode)
-			{
-				ElympicsVersionRetriever.GetVersion((version) => { _cachedSdkVersion = version; }
-				);
-			}
+			_cachedSdkVersion = ElympicsVersionRetriever.GetVersionStringFromAssembly();
 		}
 
 		private SerializedProperty GetChosenGameProperty()
@@ -93,7 +80,7 @@ namespace Elympics
 			if (_availableGames.GetValue() == null)
 				_availableGames.SetValue(new List<ElympicsGameConfig>());
 
-			
+
 
 			if (chosenGameProperty != null && chosenGameProperty.objectReferenceValue != null)
 			{
@@ -118,12 +105,7 @@ namespace Elympics
 
 		private void DrawSdkVersion()
 		{
-			var currentVersion = _elympicVersion.stringValue;
-			if (string.IsNullOrEmpty(currentVersion) || currentVersion != _cachedSdkVersion)
-			{
-				_elympicVersion.SetValue(_cachedSdkVersion);
-			}
-			EditorGUILayout.LabelField($"Elympics SDK Version: {_elympicVersion.stringValue} ");
+			EditorGUILayout.LabelField($"Elympics SDK Version: {_cachedSdkVersion} ");
 		}
 
 		private void DrawNoAvailableGamesLabel()