+++ Runtime/Config/ElympicsConfig.cs
@@ -1,18 +1,20 @@
 ﻿using System;
 using System.Collections.Generic;
-using UnityEditor;
 using UnityEngine;
+
+#if UNITY_EDITOR
+using UnityEditor;
+#endif
 
 namespace Elympics
 {
 	[CreateAssetMenu(fileName = "ElympicsConfig", menuName = "Elympics/Config")]
 	public class ElympicsConfig : ScriptableObject
 	{
-
 		public const string ELYMPICS_RESOURCES_PATH = "Assets/Resources/Elympics";
 		public const string PATH_IN_RESOURCES       = "Elympics/ElympicsConfig";
 
-		[SerializeField] private string elympicsVersion             = string.Empty;
+		private string elympicsVersion             = string.Empty;
 		[SerializeField] private string elympicsApiEndpoint         = "https://api.elympics.cc";
 		[SerializeField] private string elympicsLobbyEndpoint       = "https://lobby.elympics.cc";
 		[SerializeField] private string elympicsGameServersEndpoint = "https://gs.elympics.cc";
@@ -20,7 +22,16 @@ namespace Elympics
 		[SerializeField] internal int                      currentGame = -1;
 		[SerializeField] internal List<ElympicsGameConfig> availableGames;
 
-		internal string ElympicsVersion             => elympicsVersion;
+		internal string ElympicsVersion
+		{
+			get
+			{
+				if (string.IsNullOrEmpty(elympicsVersion))
+					elympicsVersion = ElympicsVersionRetriever.GetVersionStringFromAssembly();
+				return elympicsVersion;
+			}
+		}
+
 		internal string ElympicsApiEndpoint         => elympicsApiEndpoint;
 		internal string ElympicsLobbyEndpoint       => elympicsLobbyEndpoint;
 		internal string ElympicsGameServersEndpoint => elympicsGameServersEndpoint;
@@ -73,7 +84,7 @@ namespace Elympics
 
 		#region EditorPrefs
 
-		private const string UsernameKey = "Username";
+		private const string UsernameKey     = "Username";
 		private const string PasswordKey     = "Password";
 		private const string RefreshTokenKey = "RefreshToken";
 		private const string AuthTokenKey    = "AuthToken";