+++ Runtime/Config/ElympicsGameConfig.cs
@@ -1,9 +1,12 @@
 ﻿using System;
 using System.Collections.Generic;
 using Plugins.Elympics.Plugins.ParrelSync;
-using UnityEditor;
 using UnityEngine;
 using Object = UnityEngine.Object;
+
+#if UNITY_EDITOR
+using UnityEditor;
+#endif
 
 namespace Elympics
 {