+++ Runtime/InputsAndSnapshots/Snapshots/ElympicsSnapshot.cs
@@ -6,18 +6,30 @@ using PlayerId = System.Int32;
 
 namespace Elympics
 {
-	public class ElympicsSnapshot : ElympicsDataWithTick,IElympicsSerializable
+	public class ElympicsSnapshot : ElympicsDataWithTick, IElympicsSerializable
 	{
-		public override long                                              Tick         { get; set; }
-		public          DateTime                                          TickStartUtc { get; set; }
-		public          FactoryState                                      Factory      { get; set; }
-		public          List<KeyValuePair<int, byte[]>>                   Data         { get; set; }
-		
+		internal const int DateTimeWeight = 8;
+
+		public ElympicsSnapshot()
+		{
+		}
+
+		public ElympicsSnapshot(ElympicsSnapshot snapshot)
+		{
+			Tick = snapshot.Tick;
+			TickStartUtc = snapshot.TickStartUtc;
+			Factory = snapshot.Factory;
+			Data = snapshot.Data;
+		}
+
+		public sealed override long                            Tick         { get; set; }
+		public                 DateTime                        TickStartUtc { get; set; }
+		public                 FactoryState                    Factory      { get; set; }
+		public                 List<KeyValuePair<int, byte[]>> Data         { get; set; }
+
 		public Dictionary<PlayerId,TickToPlayerInput> TickToPlayersInputData { get; set; }
 
-
-
-		void IElympicsSerializable.Deserialize(BinaryReader br)
+		public virtual void Deserialize(BinaryReader br)
 		{
 			Tick = br.ReadInt64();
 			Factory = br.Deserialize<FactoryState>();
@@ -30,12 +42,11 @@ namespace Elympics
 				var playerToInputDataPair = br.Deserialize<TickToPlayerInput>();
 				TickToPlayersInputData.Add(tick,playerToInputDataPair);
 			}
-			var createdAt = br.ReadBytes(8);
+			var createdAt = br.ReadBytes(DateTimeWeight);
 			TickStartUtc = NtpUtils.NtpDataTimeStampToDateTime(createdAt);
 		}
 
-
-		void IElympicsSerializable.Serialize(BinaryWriter bw)
+		public virtual void Serialize(BinaryWriter bw)
 		{
 			bw.Write(Tick);
 			Factory.Serialize(bw);
@@ -46,7 +57,7 @@ namespace Elympics
 				bw.Write(keyValuePair.Key);
 				keyValuePair.Value.Serialize(bw);
 			}
-			var createdAt = new byte[8];
+			var createdAt = new byte[DateTimeWeight];
 			NtpUtils.DateTimeToNtpDataTimeStamp(TickStartUtc, createdAt);
 			bw.Write(createdAt);
 		}