+++ Runtime/InputsAndSnapshots/Snapshots/ElympicsVars/ElympicsList.cs
@@ -2,6 +2,7 @@
 using System.Collections;
 using System.Collections.Generic;
 using System.IO;
+using System.Text;
 using UnityEngine;
 
 namespace Elympics
@@ -152,6 +153,22 @@ namespace Elympics
 
 			foreach (var element in values)
 				element.Initialize(elympics);
+		}
+
+		public override string ToString()
+		{
+			var sb = new StringBuilder();
+			sb.Append($"List({Count}) [");
+			var first = true;
+			foreach (var elympicsVar in values)
+			{
+				if (!first) sb.Append(", ");
+				first = false;
+				sb.Append(elympicsVar);
+			}
+
+			sb.Append("]");
+			return sb.ToString();
 		}
 	}
 }